pyproject.toml
[project]
name = "teambuilder"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "joblib==1.5.3",
    "lightgbm>=4.6.0",
    "matplotlib==3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "scikit-learn==1.8.0",
    "scipy-stubs~=1.17.0",
]
